@@ -5,7 +5,6 @@ import (
 	"database/sql"
 	"embed"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"testing"
 	"time"
@@ -56,7 +55,7 @@ func testLoader(t *testing.T, dialect, connStr, schemaFilePath string, additiona
 		return
 	}
 
-	schema, err := ioutil.ReadFile(schemaFilePath)
+	schema, err := os.ReadFile(schemaFilePath)
 	if err != nil {
 		t.Errorf("cannot read schema file: %v", err)
 		return
@@ -79,6 +78,9 @@ func testLoader(t *testing.T, dialect, connStr, schemaFilePath string, additiona
 	}
 
 	for _, b := range batches {
+		if len(b) == 0 {
+			continue
+		}
 		if _, err = db.Exec(string(b)); err != nil {
 			t.Errorf("cannot load schema: %v", err)
 			return
@@ -454,7 +456,7 @@ func testLoader(t *testing.T, dialect, connStr, schemaFilePath string, additiona
 	})
 
 	t.Run("GenerateAndLoad", func(t *testing.T) {
-		dir, err := ioutil.TempDir(os.TempDir(), "testfixtures_test")
+		dir, err := os.MkdirTemp(os.TempDir(), "testfixtures_test")
 		if err != nil {
 			t.Errorf("cannot create temp dir: %v", err)
 			return
